[DEV-10989] Restructure `onKeyDown` handlers modules similar to normalizations
