+++ packages/slate-lists/src/onKeyDown.ts
@@ -1,10 +1,8 @@
-import { isHotkey } from 'is-hotkey';
 import type { KeyboardEvent } from 'react';
 import { Editor } from 'slate';
 import type { ReactEditor } from 'slate-react';
 
-import { canDeleteBackward, getListItemsInRange, isCursorInEmptyListItem } from './lib';
-import { decreaseDepth, increaseDepth, splitListItem } from './transformations';
+import * as OnKeyDownHandlers from './on-key-down';
 import type { ListsEditor } from './types';
 
 type CompatibleEditor = ListsEditor & ReactEditor;
@@ -28,48 +26,9 @@ export function onKeyDown(editor: CompatibleEditor, event: KeyboardEvent): boole
 }
 
 export namespace onKeyDown {
-    export function onTabIncreaseListDepth(editor: CompatibleEditor, event: KeyboardEvent) {
-        if (isHotkey('tab', event.nativeEvent)) {
-            event.preventDefault();
-            return increaseDepth(editor);
-        }
-        return false;
-    }
-
-    export function onShiftTabDecreaseListDepth(editor: CompatibleEditor, event: KeyboardEvent) {
-        if (isHotkey('shift+tab', event.nativeEvent)) {
-            event.preventDefault();
-            return decreaseDepth(editor);
-        }
-        return false;
-    }
-
-    export function onBackspaceDecreaseListDepth(editor: CompatibleEditor, event: KeyboardEvent) {
-        if (isHotkey('backspace', event.nativeEvent) && !canDeleteBackward(editor)) {
-            event.preventDefault();
-            return decreaseDepth(editor);
-        }
-        return false;
-    }
-
-    export function onEnterEscapeFromEmptyList(editor: CompatibleEditor, event: KeyboardEvent) {
-        if (isHotkey('enter', event.nativeEvent)) {
-            if (isCursorInEmptyListItem(editor)) {
-                event.preventDefault();
-                return decreaseDepth(editor);
-            }
-        }
-        return false;
-    }
-
-    export function onEnterSplitNonEmptyList(editor: CompatibleEditor, event: KeyboardEvent) {
-        if (isHotkey('enter', event.nativeEvent)) {
-            const listItemsInSelection = getListItemsInRange(editor, editor.selection);
-            if (listItemsInSelection.length > 0) {
-                event.preventDefault();
-                return splitListItem(editor);
-            }
-        }
-        return false;
-    }
+    export const onTabIncreaseListDepth = OnKeyDownHandlers.onTabIncreaseListDepth;
+    export const onShiftTabDecreaseListDepth = OnKeyDownHandlers.onShiftTabDecreaseListDepth;
+    export const onBackspaceDecreaseListDepth = OnKeyDownHandlers.onBackspaceDecreaseListDepth;
+    export const onEnterEscapeFromEmptyList = OnKeyDownHandlers.onEnterEscapeFromEmptyList;
+    export const onEnterSplitNonEmptyList = OnKeyDownHandlers.onEnterSplitNonEmptyList;
 }
